Native: optimized token activity
