+++ apps/expo/app/token/index.tsx
@@ -1,12 +1,13 @@
-import { ScreenContainer } from 'apps-expo/components/layout/ScreenContainer'
 import { Stack } from 'expo-router'
 import { IsPriceHiddenProvider } from 'app/features/home/utils/useIsPriceHidden'
-import { Paragraph, Spinner } from '@my/ui'
+import { Container, Paragraph, Spinner, useSafeAreaInsets } from '@my/ui'
 import { TokenDetails } from 'app/features/home/TokenDetails'
 import { useCoinFromTokenParam } from 'app/utils/useCoinFromTokenParam'
+import { CONTAINER_OFFSET } from 'apps-expo/components/layout/ScreenContainer'
 
 export default function TokenScreen() {
   const { coin: selectedCoin, isLoading } = useCoinFromTokenParam()
+  const insets = useSafeAreaInsets()
 
   return (
     <>
@@ -15,7 +16,17 @@ export default function TokenScreen() {
           title: 'Balance',
         }}
       />
-      <ScreenContainer>
+      <Container
+        safeAreaProps={{
+          edges: ['left', 'right'],
+          style: { flex: 1 },
+        }}
+        flex={1}
+        backgroundColor="$background"
+        overflow={'visible'}
+        paddingTop={CONTAINER_OFFSET}
+        paddingBottom={CONTAINER_OFFSET + insets.bottom}
+      >
         <IsPriceHiddenProvider>
           {(() => {
             switch (true) {
@@ -28,7 +39,7 @@ export default function TokenScreen() {
             }
           })()}
         </IsPriceHiddenProvider>
-      </ScreenContainer>
+      </Container>
     </>
   )
 }
