+++ packages/app/features/earn/rewards/screen.tsx
@@ -20,7 +20,6 @@ import { useSendEarn } from 'app/features/earn/providers/SendEarnProvider'
 import { useERC20AssetCoin } from 'app/features/earn/params'
 import { useSendEarnClaimRewardsCalls } from 'app/features/earn/rewards/hooks'
 import { TokenActivityRow } from 'app/features/home/TokenActivityRow'
-import { TokenDetailsMarketData } from 'app/features/home/TokenDetails'
 import { assert } from 'app/utils/assert'
 import { formatCoinAmount } from 'app/utils/formatCoinAmount'
 import { useSendAccount } from 'app/utils/send-accounts'
@@ -33,6 +32,7 @@ import { SectionList } from 'react-native'
 import { formatUnits, withRetry } from 'viem'
 import { useChainId } from 'wagmi'
 import { useEarnRewardsActivityFeed } from './hooks'
+import { TokenDetailsMarketData } from 'app/features/home/TokenDetailsHeader'
 
 const log = debug('app:features:earn:rewards')
