+++ packages/app/features/home/TokenActivity.tsx
@@ -1,11 +1,12 @@
-import { H4, Paragraph, Spinner, YStack } from '@my/ui'
+import { Paragraph, Spinner, YStack, H4 } from '@my/ui'
 import type { CoinWithBalance } from 'app/data/coins'
 import { hexToBytea } from 'app/utils/hexToBytea'
 import { toNiceError } from 'app/utils/toNiceError'
 import { ActivityDetails } from '../activity/ActivityDetails'
-import { TokenActivityFeed } from './TokenActivityFeed'
+import TokenActivityFeed from './TokenActivityFeed'
 import { useTokenActivityFeed } from './utils/useTokenActivityFeed'
 import { useActivityDetails } from 'app/provider/activity-details'
+import { Platform } from 'react-native'
 
 export const TokenActivity = ({ coin }: { coin: CoinWithBalance }) => {
   const { isOpen, selectActivity } = useActivityDetails()
@@ -31,20 +32,30 @@ export const TokenActivity = ({ coin }: { coin: CoinWithBalance }) => {
   if (isOpen) {
     return <ActivityDetails />
   }
-  return (
-    <YStack gap={'$3'}>
-      <H4 fontWeight={'600'} size={'$7'}>
-        {!pages || !pages[0]?.length ? 'No Activity' : 'Activity'}
-      </H4>
-      <TokenActivityFeed
-        testID="TokenActivityFeed"
-        tokenActivityFeedQuery={tokenActivityFeedQuery}
-        onActivityPress={selectActivity}
-        $gtLg={{
-          p: '$3.5',
-        }}
-        p="$2"
-      />
-    </YStack>
+
+  const content = (
+    <TokenActivityFeed
+      testID="TokenActivityFeed"
+      tokenActivityFeedQuery={tokenActivityFeedQuery}
+      onActivityPress={selectActivity}
+      coin={coin}
+      $gtLg={{
+        p: '$3.5',
+      }}
+      p="$2"
+    />
   )
+
+  if (Platform.OS === 'web') {
+    return (
+      <YStack gap={'$3'}>
+        <H4 fontWeight={'600'} size={'$7'}>
+          {!pages || !pages[0]?.length ? 'No Activity' : 'Activity'}
+        </H4>
+        {content}
+      </YStack>
+    )
+  }
+
+  return content
 }
