+++ packages/ui/src/components/index.ts
@@ -19,7 +19,8 @@ export * from './DatePicker'
 export * from './SafeArea'
 export * from './FadeCard'
 export * from './LinkBanner'
-export { RecyclerList } from './RecyclerList.web'
+export * from './RecyclerList.web'
+export * from './RecyclerList.native'
 export * from './PendingIndicatorBar'
 export * from './Card'
 export * from './BlurStack'
